moved tooltip to utils.ts and removed the file
